feat: add more automation
The Release Please task has been added.
The `base` and `all` aliases will no longer execute the package task.

The TypeScript out dir is now in the prettierignore and eslintignore.
The module is now CommonJS for the tsconfig.
Rimraf is now installed and used with TypeScript.
The out dir is added to the `files` field of the package
when TypeScript is used.

Release-As: 2.0.1